+++ internal/controllers/rpc_test.go
@@ -4,16 +4,20 @@ import (
 	"bytes"
 	"encoding/json"
 	"errors"
+	"github.com/NodeFactoryIo/vedran/internal/configuration"
 	"github.com/NodeFactoryIo/vedran/internal/models"
 	"github.com/NodeFactoryIo/vedran/internal/repositories"
 	"github.com/NodeFactoryIo/vedran/internal/rpc"
 	actionMocks "github.com/NodeFactoryIo/vedran/mocks/actions"
+	tunnelMocks "github.com/NodeFactoryIo/vedran/mocks/http-tunnel/server"
 	repoMocks "github.com/NodeFactoryIo/vedran/mocks/repositories"
 	"github.com/stretchr/testify/mock"
 	"io"
 	"net/http"
 	"net/http/httptest"
+	"net/url"
 	"reflect"
+	"strconv"
 	"testing"
 )
 
@@ -37,20 +41,26 @@ func TestApiController_RPCHandler(t *testing.T) {
 	setup()
 	defer teardown()
 
+	poolerMock := &tunnelMocks.Pooler{}
+	configuration.Config.PortPool = poolerMock
+
 	nodeRepoMock := repoMocks.NodeRepository{}
 	pingRepoMock := repoMocks.PingRepository{}
 	metricsRepoMock := repoMocks.MetricsRepository{}
 	recordRepoMock := repoMocks.RecordRepository{}
+	recordRepoMock.On("Save", mock.Anything).Return(nil)
+
 	actionsMockObject := new(actionMocks.Actions)
 	actionsMockObject.On("PenalizeNode", mock.Anything, mock.Anything).Return()
 	actionsMockObject.On("RewardNode", mock.Anything, mock.Anything).Return()
-	recordRepoMock.On("Save", mock.Anything).Return(nil)
+
 	apiController := NewApiController(false, repositories.Repos{
 		NodeRepo:    &nodeRepoMock,
 		PingRepo:    &pingRepoMock,
 		MetricsRepo: &metricsRepoMock,
 		RecordRepo:  &recordRepoMock,
 	}, actionsMockObject)
+
 	handler := http.HandlerFunc(apiController.RPCHandler)
 
 	tests := []struct {
@@ -61,6 +71,18 @@ func TestApiController_RPCHandler(t *testing.T) {
 		handleFunc  handleFnMock
 	}{
 		{
+			name:       "Returns response if node returnes valid rpc response",
+			rpcRequest: `{"jsonrpc": "2.0", "id": 1, "method": "system"}`,
+			rpcResponse: rpc.RPCResponse{
+				ID:      1,
+				JSONRPC: "2.0",
+				Error:   nil,
+			},
+			nodes: []models.Node{{ID: "test-id"}},
+			handleFunc: func(w http.ResponseWriter, r *http.Request) {
+				_, _ = io.WriteString(w, `{"id": 1, "jsonrpc": "2.0"}`)
+			}},
+		{
 			name:       "Returns parse error if json invalid",
 			rpcRequest: `INVALID`,
 			rpcResponse: rpc.RPCResponse{
@@ -68,7 +90,9 @@ func TestApiController_RPCHandler(t *testing.T) {
 				JSONRPC: "2.0",
 				Error:   &rpc.RPCError{Code: -32700, Message: "Parse error"},
 			},
-		},
+			handleFunc: func(w http.ResponseWriter, r *http.Request) {
+				_, _ = io.WriteString(w, `{"id": 1, "jsonrpc": "2.0"}`)
+			}},
 		{
 			name:       "Returns server error if no available nodes",
 			rpcRequest: `{"jsonrpc": "2.0", "id": 1, "method": "system"}`,
@@ -78,7 +102,9 @@ func TestApiController_RPCHandler(t *testing.T) {
 				Error:   &rpc.RPCError{Code: -32603, Message: "No available nodes"},
 			},
 			nodes: []models.Node{},
-		},
+			handleFunc: func(w http.ResponseWriter, r *http.Request) {
+				_, _ = io.WriteString(w, `{"id": 1, "jsonrpc": "2.0"}`)
+			}},
 		{
 			name:       "Returns server error if all nodes return invalid rpc response",
 			rpcRequest: `{"jsonrpc": "2.0", "id": 1, "method": "system"}`,
@@ -87,17 +113,7 @@ func TestApiController_RPCHandler(t *testing.T) {
 				JSONRPC: "2.0",
 				Error:   &rpc.RPCError{Code: -32603, Message: "Internal Server Error"},
 			},
-			nodes: []models.Node{{ID: "test-id", NodeUrl: "invalid"}},
-		},
-		{
-			name:       "Returns response if node returnes valid rpc response",
-			rpcRequest: `{"jsonrpc": "2.0", "id": 1, "method": "system"}`,
-			rpcResponse: rpc.RPCResponse{
-				ID:      1,
-				JSONRPC: "2.0",
-				Error:   nil,
-			},
-			nodes: []models.Node{{ID: "test-id", NodeUrl: "valid"}},
+			nodes: []models.Node{{ID: "test-id"}},
 			handleFunc: func(w http.ResponseWriter, r *http.Request) {
 				_, _ = io.WriteString(w, `{"id": 1, "jsonrpc": "2.0"}`)
 			},
@@ -108,18 +124,13 @@ func TestApiController_RPCHandler(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			setup()
 
-			if test.handleFunc != nil {
-				mux.HandleFunc("/", test.handleFunc)
-			}
-			if len(test.nodes) > 0 && test.nodes[0].NodeUrl == "valid" {
-				test.nodes[0].NodeUrl = server.URL
-			} else if len(test.nodes) > 0 {
-				test.nodes[0].NodeUrl = "INVALID"
-			}
+			serverURL, _ := url.Parse(server.URL)
+			port, _ := strconv.Atoi(serverURL.Port())
+			poolerMock.On("GetPort", mock.Anything).Once().Return(port, nil)
+
+			mux.HandleFunc("/", test.handleFunc)
 
 			nodeRepoMock.On("GetActiveNodes", mock.Anything).Return(&test.nodes, nil)
-			nodeRepoMock.On("RewardNode", mock.Anything).Return()
-			nodeRepoMock.On("PenalizeNode", mock.Anything).Return()
 
 			req, _ := http.NewRequest("POST", "/", bytes.NewReader([]byte(test.rpcRequest)))
 			rr := httptest.NewRecorder()
@@ -143,20 +154,26 @@ func TestApiController_BatchRPCHandler(t *testing.T) {
 	setup()
 	defer teardown()
 
+	poolerMock := &tunnelMocks.Pooler{}
+	configuration.Config.PortPool = poolerMock
+
 	nodeRepoMock := repoMocks.NodeRepository{}
 	pingRepoMock := repoMocks.PingRepository{}
 	metricsRepoMock := repoMocks.MetricsRepository{}
 	recordRepoMock := repoMocks.RecordRepository{}
 	recordRepoMock.On("Save", mock.Anything).Return(nil)
+
 	actionsMockObject := new(actionMocks.Actions)
 	actionsMockObject.On("PenalizeNode", mock.Anything, mock.Anything).Return()
 	actionsMockObject.On("RewardNode", mock.Anything, mock.Anything).Return()
+
 	apiController := NewApiController(false, repositories.Repos{
 		NodeRepo:    &nodeRepoMock,
 		PingRepo:    &pingRepoMock,
 		MetricsRepo: &metricsRepoMock,
 		RecordRepo:  &recordRepoMock,
 	}, actionsMockObject)
+
 	handler := http.HandlerFunc(apiController.RPCHandler)
 
 	tests := []struct {
@@ -175,24 +192,22 @@ func TestApiController_BatchRPCHandler(t *testing.T) {
 					JSONRPC: "2.0",
 					Error:   &rpc.RPCError{Code: -32603, Message: "Internal Server Error"}},
 			},
-			nodes: []models.Node{{ID: "test-id", NodeUrl: "invalid"}}},
+			nodes: []models.Node{{ID: "test-id"}}},
 	}
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			setup()
 
+			serverURL, _ := url.Parse(server.URL)
+			port, _ := strconv.Atoi(serverURL.Port())
+			poolerMock.On("GetPort", mock.Anything).Once().Return(port, nil)
+
 			if test.handleFunc != nil {
 				mux.HandleFunc("/", test.handleFunc)
 			}
-			if len(test.nodes) > 0 && test.nodes[0].NodeUrl == "valid" {
-				test.nodes[0].NodeUrl = server.URL
-			} else if len(test.nodes) > 0 {
-				test.nodes[0].NodeUrl = "INVALID"
-			}
 
 			nodeRepoMock.On("GetActiveNodes", mock.Anything).Return(&test.nodes, nil)
-			nodeRepoMock.On("PenalizeNode", mock.Anything).Return()
 
 			req, _ := http.NewRequest("POST", "/", bytes.NewReader([]byte(test.rpcRequest)))
 			rr := httptest.NewRecorder()
@@ -237,15 +252,18 @@ func TestApiController_RPCHandler_InvalidBody(t *testing.T) {
 	pingRepoMock := repoMocks.PingRepository{}
 	metricsRepoMock := repoMocks.MetricsRepository{}
 	recordRepoMock := repoMocks.RecordRepository{}
+
 	actionsMockObject := new(actionMocks.Actions)
 	actionsMockObject.On("PenalizeNode", mock.Anything, mock.Anything).Return()
 	actionsMockObject.On("RewardNode", mock.Anything, mock.Anything).Return()
+
 	apiController := NewApiController(false, repositories.Repos{
 		NodeRepo:    &nodeRepoMock,
 		PingRepo:    &pingRepoMock,
 		MetricsRepo: &metricsRepoMock,
 		RecordRepo:  &recordRepoMock,
 	}, actionsMockObject)
+
 	handler := http.HandlerFunc(apiController.RPCHandler)
 
 	for _, test := range tests {