+++ internal/models/node.go
@@ -3,7 +3,6 @@ package models
 type Node struct {
 	ID            string `storm:"id"`
 	ConfigHash    string
-	NodeUrl       string
 	PayoutAddress string
 	Token         string
 	Cooldown      int