+++ CHANGELOG.md
@@ -8,8 +8,10 @@
 - Implement tunnel server in load balancer [\#63](https://github.com/NodeFactoryIo/vedran/pull/63) ([MakMuftic](https://github.com/MakMuftic))
 
 ### Fix
+- Fix panic if no nodes in database after startup [\#62](https://github.com/NodeFactoryIo/vedran/pull/62) ([mpetrun5](https://github.com/mpetrun5))
 
 ### Changed
+- Use port from tunnel map for calling node [\#65](https://github.com/NodeFactoryIo/vedran/pull/65) ([mpetrun5](https://github.com/mpetrun5))
 
 ## [v0.1.1]((https://github.com/NodeFactoryIo/vedran/tree/v0.1.1))